Split GCToEEInterface definition out of gc.h into its own header
